Pages Set Up
In a file-based, non-local-host way, this is the set up to the website. This commit adds all of the views, in working order this time, and allows for a user to switch between them easily. All JS was removed as this seems to be an easier way. There are 5 html files for the five pages of the website, other than sign-in.