Formatting, console log, & read me updates
formatting: I added some css & reordered some of the elements to make reading the devpage more easy & appealing

console log: I added the console logs so that it was easier to demonstrate the routes in the video recording

readme: This should be the complete readme up to this point in the project. It should include information on how to interact with the API & some information about the code itself.